add vegetable fragment
