+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w7/main/VegetableFragment.kt
@@ -1,0 +1,23 @@
+package com.asiantech.intern20summer1.w7.main
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.asiantech.intern20summer1.R
+
+class VegetableFragment : Fragment(){
+
+    companion object{
+        internal fun newInstance() = VegetableFragment()
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_farm_vegetable,container,false)
+    }
+}
