--- app/src/at-huybui/java/com/asiantech/intern20summer1/w7/main/recycalview/RecycalItem.kt
@@ -1,3 +1,0 @@
-package com.asiantech.intern20summer1.w7.main.recycalview
-
-data class RecycalItem(val name: String = "", val infomation: String = "", val image: Int = 0)
