+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w7/main/recyclerview/RecyclerItem.kt
@@ -1,0 +1,3 @@
+package com.asiantech.intern20summer1.w7.main.recyclerview
+
+data class RecyclerItem(val name: String = "", val infomation: String = "", val image: Int = 0)
