+++ app/src/at-huybui/res/layout/activity_main_farm.xml
@@ -1,33 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/containerMain"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.asiantech.intern20summer1.w7.main.MainFarmActivity">
-    <androidx.cardview.widget.CardView
-        android:id="@+id/cardViewAvatarHome"
-        android:layout_width="@dimen/home_avatar_size"
-        android:layout_height="@dimen/home_avatar_size"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_gravity="center"
-        android:layout_marginTop="@dimen/w7_avatar_margin_top"
-        app:cardCornerRadius="@dimen/home_avatar_corner_radius">
-
-        <ImageView
-            android:id="@+id/imgAvatarHome"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:scaleType="centerCrop"
-            android:src="@drawable/img_avatar_default" />
-    </androidx.cardview.widget.CardView>
-    <TextView
-        android:id="@+id/tvTest"
-        android:layout_below="@id/cardViewAvatarHome"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="30dp"
-        android:layout_centerInParent="true" />
-
-</RelativeLayout>
+    android:background="@color/background_white" />
