+++ app/src/at-huybui/res/values/dimens.xml
@@ -73,6 +73,8 @@
     <dimen name="w7_card_view_radius">16dp</dimen>
     <dimen name="w7_card_view_margin">4dp</dimen>
     <dimen name="w7_home_image_vegetable_size">80dp</dimen>
-    <dimen name="w8_home_name_vegetable_size">22sp</dimen>
-    <dimen name="w8_home_name_vegetable_margin_top_left">20dp</dimen>
+    <dimen name="w7_home_name_vegetable_size">22sp</dimen>
+    <dimen name="w7_home_name_vegetable_margin_top_left">20dp</dimen>
+    <dimen name="w7_home_information_vegetable_margin_top">52dp</dimen>
+    <dimen name="w7_home_information_vegetable_margin_left">28dp</dimen>
 </resources>
